dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile dev-requirements.in
#
appdirs==1.4.4
    # via
    #   black
    #   virtualenv
astroid==2.5.2
    # via pylint
attrs==20.3.0
    # via pytest
black==20.8b1
    # via -r dev-requirements.in
cfgv==3.2.0
    # via pre-commit
click==7.1.2
    # via black
coverage==5.5
    # via -r dev-requirements.in
distlib==0.3.1
    # via virtualenv
filelock==3.0.12
    # via virtualenv
identify==2.2.2
    # via pre-commit
iniconfig==1.1.1
    # via pytest
isort==5.8.0
    # via pylint
lazy-object-proxy==1.6.0
    # via astroid
mccabe==0.6.1
    # via pylint
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
mypy==0.812
    # via -r dev-requirements.in
nodeenv==1.5.0
    # via pre-commit
packaging==20.9
    # via pytest
pathspec==0.8.1
    # via black
pluggy==0.13.1
    # via pytest
pre-commit==2.11.1
    # via -r dev-requirements.in
py==1.10.0
    # via pytest
pylint-django==2.4.2
    # via -r dev-requirements.in
pylint-plugin-utils==0.6
    # via pylint-django
pylint==2.7.4
    # via
    #   -r dev-requirements.in
    #   pylint-django
    #   pylint-plugin-utils
pyparsing==2.4.7
    # via packaging
pytest-django==4.1.0
    # via -r dev-requirements.in
pytest==6.2.3
    # via pytest-django
pyyaml==5.4.1
    # via pre-commit
regex==2021.4.4
    # via black
six==1.15.0
    # via virtualenv
toml==0.10.2
    # via
    #   black
    #   pre-commit
    #   pylint
    #   pytest
typed-ast==1.4.2
    # via
    #   black
    #   mypy
typing-extensions==3.7.4.3
    # via
    #   black
    #   mypy
virtualenv==20.4.3
    # via pre-commit
wrapt==1.12.1
    # via astroid
